# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Intercultural Experience Platform API",
        "description": "API completa para la plataforma de intercambios culturales. Incluye autenticación, gestión de programas, aplicaciones, sistema de puntos y más.",
        "contact": {
            "name": "IE Development Team",
            "email": "developers@ie.org.py"
        },
        "license": {
            "name": "Proprietary",
            "url": "https://ie.org.py/terms"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://ie.org.py/api",
            "description": "Servidor de Producción"
        },
        {
            "url": "https://staging.ie.org.py/api",
            "description": "Servidor de Staging"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor de Desarrollo Local"
        }
    ],
    "paths": {
...
        "/job-offers": {
            "get": {
                "tags": [
                    "Job Offers"
                ],
                "summary": "Listar ofertas laborales disponibles",
                "description": "Obtiene todas las ofertas laborales disponibles con filtros opcionales",
                "operationId": "7535dea4abe3bcc7b6c1882cb0d24ef4",
                "parameters": [
                    {
                        "name": "city",
                        "in": "query",
                        "description": "Filtrar por ciudad",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Miami"
                        }
                    },
                    {
                        "name": "state",
                        "in": "query",
                        "description": "Filtrar por estado",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Florida"
                        }
                    },
                    {
                        "name": "english_level",
                        "in": "query",
                        "description": "Filtrar por nivel de inglés requerido",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "A2",
                                "B1",
                                "B1+",
                                "B2",
                                "C1",
                                "C2"
                            ]
                        }
                    },
                    {
                        "name": "gender",
                        "in": "query",
                        "description": "Filtrar por género requerido",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "male",
                                "female",
                                "any"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Resultados por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de ofertas obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "job_offer_id": {
                                                                "type": "string",
                                                                "example": "JO-2025-001"
                                                            },
                                                            "position": {
                                                                "type": "string",
                                                                "example": "Lifeguard"
                                                            },
                                                            "city": {
                                                                "type": "string",
                                                                "example": "Miami"
                                                            },
                                                            "state": {
                                                                "type": "string",
                                                                "example": "Florida"
                                                            },
                                                            "salary_min": {
                                                                "type": "number",
                                                                "example": 2000
                                                            },
                                                            "salary_max": {
                                                                "type": "number",
                                                                "example": 2500
                                                            },
                                                            "available_slots": {
                                                                "type": "integer",
                                                                "example": 5
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/job-offers/{id}": {
            "get": {
                "tags": [
                    "Job Offers"
                ],
                "summary": "Obtener detalles de una oferta laboral",
                "description": "Obtiene los detalles completos de una oferta laboral específica",
                "operationId": "3c0768a152e77f135ae6e20d9efd2608",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la oferta laboral",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la oferta obtenidos exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "job_offer_id": {
                                                    "type": "string",
                                                    "example": "JO-2025-001"
                                                },
                                                "position": {
                                                    "type": "string",
                                                    "example": "Lifeguard"
                                                },
                                                "description": {
                                                    "type": "string"
                                                },
                                                "city": {
                                                    "type": "string",
                                                    "example": "Miami"
                                                },
                                                "state": {
                                                    "type": "string",
                                                    "example": "Florida"
                                                },
                                                "salary_min": {
                                                    "type": "number",
                                                    "example": 2000
                                                },
                                                "salary_max": {
                                                    "type": "number",
                                                    "example": 2500
                                                },
                                                "hours_per_week": {
                                                    "type": "integer",
                                                    "example": 40
                                                },
                                                "housing_type": {
                                                    "type": "string",
                                                    "example": "provided"
                                                },
                                                "total_slots": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "available_slots": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "required_english_level": {
                                                    "type": "string",
                                                    "example": "B2"
                                                },
                                                "required_gender": {
                                                    "type": "string",
                                                    "example": "any"
                                                },
                                                "start_date": {
                                                    "type": "string",
                                                    "format": "date"
                                                },
                                                "end_date": {
                                                    "type": "string",
                                                    "format": "date"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Oferta no encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/job-offers/recommended": {
            "get": {
                "tags": [
                    "Job Offers"
                ],
                "summary": "Obtener ofertas recomendadas",
                "description": "Obtiene ofertas laborales recomendadas basadas en el perfil del usuario (nivel de inglés y género)",
                "operationId": "f62f97b3923c696972d9276d7eefc776",
                "responses": {
                    "200": {
                        "description": "Ofertas recomendadas obtenidas exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "position": {
                                                        "type": "string"
                                                    },
                                                    "city": {
                                                        "type": "string"
                                                    },
                                                    "match_score": {
                                                        "description": "Porcentaje de compatibilidad",
                                                        "type": "integer",
                                                        "example": 95
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}